refactor: modularize question preview components by moving them into individual files
