+++ assets/src/js/v3/entries/course-builder/components/modals/QuestionPreviewModal.tsx
@@ -14,7 +14,19 @@ import { tutorConfig } from '@TutorShared/config/config';
 import { borderRadius, Breakpoint, colorTokens, spacing } from '@TutorShared/config/styles';
 import Show from '@TutorShared/controls/Show';
 import { styleUtils } from '@TutorShared/utils/style-utils';
-import { type QuizQuestion, type QuizQuestionOption, type QuizQuestionType } from '@TutorShared/utils/types';
+import { type QuizQuestion, type QuizQuestionType } from '@TutorShared/utils/types';
+
+import DrawImagePreview from './question-previews/DrawImagePreview';
+import FillInTheBlankPreview from './question-previews/FillInTheBlankPreview';
+import ImageAnsweringPreview from './question-previews/ImageAnsweringPreview';
+import MatchingPreview from './question-previews/MatchingPreview';
+import MultipleChoicePreview from './question-previews/MultipleChoicePreview';
+import OpenEndedPreview from './question-previews/OpenEndedPreview';
+import OrderingPreview from './question-previews/OrderingPreview';
+import PinImagePreview from './question-previews/PinImagePreview';
+import ScalePreview from './question-previews/ScalePreview';
+import TrueFalsePreview from './question-previews/TrueFalsePreview';
+import UnsupportedPreview from './question-previews/UnsupportedPreview';
 
 interface QuestionPreviewModalProps extends ModalProps {
   question: QuizQuestion;
@@ -324,369 +336,6 @@ const renderQuestionPreview = (question: QuizQuestion) => {
     default:
       return <UnsupportedPreview />;
   }
-};
-
-const TrueFalsePreview = ({ answers }: { answers: QuizQuestionOption[] }) => (
-  <div className="tutor-quiz-question-options">
-    {answers.map((answer) => (
-      <label key={answer.answer_id} className="tutor-quiz-question-option">
-        <SVGIcon name={__('True', 'tutor') === answer.answer_title ? 'check2' : 'cross'} width={20} height={20} />
-        {answer.answer_title}
-      </label>
-    ))}
-  </div>
-);
-
-const MultipleChoicePreview = ({
-  answers,
-  hasMultipleCorrectAnswer,
-}: {
-  answers: QuizQuestionOption[];
-  hasMultipleCorrectAnswer: boolean;
-}) => (
-  <div className="tutor-quiz-question-options">
-    {answers.map((answer, index) => {
-      const hasImage = !!answer.image_url;
-
-      return (
-        <label key={answer.answer_id || index} className="tutor-quiz-question-option">
-          <Show
-            when={!hasImage}
-            fallback={
-              <>
-                <img src={answer.image_url} alt={answer.answer_title} />
-                <div data-title>{answer.answer_title}</div>
-              </>
-            }
-          >
-            <div className="tutor-input-field">
-              <div className="tutor-input-wrapper">
-                <input
-                  type={hasMultipleCorrectAnswer ? 'checkbox' : 'radio'}
-                  className={hasMultipleCorrectAnswer ? 'tutor-checkbox' : 'tutor-radio'}
-                  disabled
-                />
-                <label>{answer.answer_title}</label>
-              </div>
-            </div>
-          </Show>
-        </label>
-      );
-    })}
-  </div>
-);
-
-const OpenEndedPreview = ({ questionType }: { questionType: QuizQuestionType }) => (
-  <div className="tutor-quiz-question-options">
-    <textarea
-      className="tutor-form-textarea tutor-quiz-question-input"
-      placeholder={
-        questionType === 'short_answer'
-          ? __('Type your short answer here', 'tutor')
-          : __('Type your answer here', 'tutor')
-      }
-      disabled
-      rows={5}
-    />
-  </div>
-);
-
-const FillInTheBlankPreview = ({ answers }: { answers: QuizQuestionOption[] }) => (
-  <div className="tutor-quiz-question-options">
-    {answers.map((answer, index) => (
-      <div key={answer.answer_id || index} className="tutor-quiz-question-option">
-        {renderFillInBlankText(answer, index)}
-      </div>
-    ))}
-  </div>
-);
-
-const MatchingPreview = ({ answers, isImageMatching }: { answers: QuizQuestionOption[]; isImageMatching: boolean }) => {
-  const draggableAnswers = [...answers].sort((first, second) => first.answer_order - second.answer_order);
-
-  return (
-    <div className="tutor-flex tutor-flex-column tutor-gap-7 tutor-sm-gap-5">
-      <div className="tutor-quiz-question-options" data-image-matching={isImageMatching ? '1' : '0'}>
-        {answers.map((answer, index) => (
-          <div key={answer.answer_id || index} className="tutor-quiz-question-option">
-            <Show
-              when={isImageMatching && answer.image_url}
-              fallback={
-                <div data-title>
-                  <div className="tutor-quiz-question-option-number">{answer.answer_order || index + 1}</div>
-                  {answer.answer_title}
-                </div>
-              }
-            >
-              <img src={answer.image_url} alt={answer.answer_title} />
-            </Show>
-            <div className="tutor-quiz-question-option-drop-zone" data-drop-placeholder-text={__('Drop here', 'tutor')}>
-              <span data-drop-placeholder className="tutor-text-subdued">
-                {__('Drop here', 'tutor')}
-              </span>
-            </div>
-          </div>
-        ))}
-      </div>
-
-      <div className="tutor-quiz-question-draggable">
-        <div className="tutor-quiz-question-draggable-header">
-          <SVGIcon name="drag" width={20} height={20} />
-          <span className="tutor-text-small tutor-font-medium">{__('Drag from here', 'tutor')}</span>
-        </div>
-        <div className="tutor-quiz-question-options">
-          {draggableAnswers.map((answer, index) => (
-            <div
-              key={answer.answer_id || index}
-              className="tutor-quiz-question-option"
-              data-option="draggable"
-              data-id={answer.answer_id}
-            >
-              <div data-title>
-                <Show when={isImageMatching} fallback={answer.answer_two_gap_match}>
-                  {answer.answer_title}
-                </Show>
-              </div>
-              <button type="button" data-grab-handle disabled>
-                <SVGIcon name="grabHandle" width={24} height={24} />
-              </button>
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const ImageAnsweringPreview = ({ answers }: { answers: QuizQuestionOption[] }) => (
-  <div className="tutor-quiz-question-options">
-    {answers.map((answer, index) => (
-      <div key={answer.answer_id || index} className="tutor-quiz-question-option">
-        <Show when={answer.image_url}>
-          <img src={answer.image_url} alt={answer.answer_title} />
-        </Show>
-        <input
-          type="text"
-          className="tutor-form-control tutor-quiz-question-input"
-          placeholder={__('Write your answer here', 'tutor')}
-          disabled
-        />
-      </div>
-    ))}
-  </div>
-);
-
-const OrderingPreview = ({ answers }: { answers: QuizQuestionOption[] }) => (
-  <div className="tutor-quiz-question-options">
-    {answers.map((answer, index) => (
-      <div
-        key={answer.answer_id || index}
-        className="tutor-quiz-question-option"
-        data-option="draggable"
-        data-id={answer.answer_id}
-      >
-        <div data-option-order>{answer.answer_order || index + 1}</div>
-        <div data-title>
-          {answer.image_url ? <img src={answer.image_url} alt={answer.answer_title} /> : null}
-          {answer.answer_title}
-        </div>
-
-        <button type="button" data-grab-handle disabled>
-          <SVGIcon name="grabHandle" width={24} height={24} />
-        </button>
-      </div>
-    ))}
-  </div>
-);
-
-const UnsupportedPreview = () => (
-  <div className="tutor-quiz-question-options">
-    <div className="tutor-quiz-question-option">
-      {__('Preview is not available for this question type yet.', 'tutor')}
-    </div>
-  </div>
-);
-
-const PinImagePreview = ({ answers }: { answers: QuizQuestionOption[] }) => {
-  const imageUrl = answers[0]?.image_url;
-  if (!imageUrl) {
-    return (
-      <div className="tutor-quiz-question-options">
-        <p className="tutor-fs-7 tutor-color-secondary">
-          {__('No background image configured for this Pin question.', 'tutor')}
-        </p>
-      </div>
-    );
-  }
-  return (
-    <div className="tutor-quiz-question-options tutor-pin-image-question">
-      <div className="tutor-pin-image-wrapper">
-        <img src={imageUrl} alt={__('Pin on image question', 'tutor')} />
-        <span className="tutor-pin-image-marker" aria-hidden="true" />
-      </div>
-    </div>
-  );
-};
-
-const DrawImagePreview = ({ answers }: { answers: QuizQuestionOption[] }) => {
-  const imageUrl = answers[0]?.image_url;
-  if (!imageUrl) {
-    return (
-      <div className="tutor-quiz-question-options">
-        <p className="tutor-fs-7 tutor-color-secondary">
-          {__('No background image configured for this Mark in the Image question.', 'tutor')}
-        </p>
-      </div>
-    );
-  }
-  return (
-    <div className="tutor-quiz-question-options tutor-draw-image-question">
-      <div className="tutor-draw-image-wrapper">
-        <img src={imageUrl} alt={__('Draw on image question', 'tutor')} />
-        <canvas className="tutor-draw-image-canvas" />
-      </div>
-    </div>
-  );
-};
-
-const SCALE_SCRIPT_ATTR = 'data-tutor-scale-preview-script';
-
-const ScalePreview = ({ answers }: { answers: QuizQuestionOption[] }) => {
-  const wrapperRef = useRef<HTMLDivElement>(null);
-
-  const configRaw = answers[0]?.answer_two_gap_match ?? '';
-  let rawConfig: Record<string, unknown> = {};
-  try {
-    const parsed = JSON.parse(configRaw) as { config?: Record<string, unknown> };
-    rawConfig = parsed?.config ?? {};
-  } catch {
-    // use defaults
-  }
-
-  const min = (rawConfig.min as number) ?? 0;
-  const max = (rawConfig.max as number) ?? 100;
-  const step = (rawConfig.step as number) ?? 1;
-  const defaultValue = (rawConfig.defaultValue as number) ?? (min + max) / 2;
-  const pxPerUnit = (rawConfig.pxPerUnit as number) ?? 10;
-  const labelEvery = (rawConfig.labelEvery as number) ?? Math.max(1, (max - min) / 10);
-  const minorTickEvery = (rawConfig.minorTickEvery as number) ?? Math.max(1, (max - min) / 50);
-  const precision = (rawConfig.precision as number) ?? (step < 1 ? 2 : 0);
-
-  // Unique IDs matching scale-question.js selectors
-  const qId = answers[0]?.answer_id ?? 'preview';
-  const scaleConfig = JSON.stringify({
-    min,
-    max,
-    step,
-    defaultValue,
-    pxPerUnit,
-    labelEvery,
-    minorTickEvery,
-    precision,
-  });
-
-  useEffect(() => {
-    const wrapper = wrapperRef.current;
-    if (!wrapper) return;
-
-    const doc = wrapper.ownerDocument;
-    if (!doc || doc === document) return;
-
-    if (doc.head.querySelector(`[${SCALE_SCRIPT_ATTR}]`)) {
-      // Script already present — reset the wrapper init flag so it re-initializes
-      wrapper.removeAttribute('data-tutor-scale-init');
-      // Re-trigger initAll via a tiny inline script on the iframe window
-      const reinit = doc.createElement('script');
-      reinit.textContent = 'if(window._tutorScaleInitAll){window._tutorScaleInitAll();}';
-      doc.head.appendChild(reinit);
-      return;
-    }
-
-    const siteUrl = tutorConfig.site_url.replace(/\/$/, '');
-    const script = doc.createElement('script');
-    script.setAttribute(SCALE_SCRIPT_ATTR, '1');
-    // Patch initAll onto window so we can call it on re-renders
-    script.textContent = `
-      (function(){
-        var scriptEl = document.createElement('script');
-        scriptEl.src = '${siteUrl}/wp-content/plugins/tutor-pro/assets/js/scale-question.js';
-        scriptEl.onload = function(){
-          // Expose initAll for re-use on subsequent renders
-          if(typeof initAll === 'function'){ window._tutorScaleInitAll = function(){ initAll(false); }; }
-        };
-        document.head.appendChild(scriptEl);
-      })();
-    `;
-    doc.head.appendChild(script);
-  }, []);
-
-  return (
-    <div
-      ref={wrapperRef}
-      id={`tutor-scale-question-${qId}`}
-      className="tutor-quiz-question-options tutor-scale-question"
-      data-question-type="scale"
-      data-question-id={String(qId)}
-      data-scale-config={scaleConfig}
-    >
-      <div className="tutor-scale-slider-wrapper">
-        <div className="tutor-scale-bubble" id={`tutor-scale-bubble-${qId}`}>
-          <div className="tutor-scale-bubble-value">{defaultValue}</div>
-          <div className="tutor-scale-bubble-pointer" />
-        </div>
-        <div
-          className="tutor-scale-container"
-          id={`tutor-scale-container-${qId}`}
-          aria-label={__('Interactive scale: drag to select your answer value.', 'tutor')}
-        >
-          <div className="tutor-scale" id={`tutor-scale-${qId}`}>
-            {/* Ticks are generated by scale-question.js */}
-          </div>
-        </div>
-        <div className="tutor-scale-instructions">
-          <div className="tutor-scale-hand-icon" aria-hidden="true">
-            <SVGIcon name="handSwipeRight" width={20} height={20} />
-          </div>
-          <span className="tutor-scale-instruction-text">
-            {__('Move the scale left or right to set value', 'tutor')}
-          </span>
-        </div>
-      </div>
-      <input
-        type="hidden"
-        id={`tutor-scale-value-${qId}`}
-        name={`preview[answers][scale][value]`}
-        defaultValue=""
-        readOnly
-      />
-    </div>
-  );
-};
-
-const renderFillInBlankText = (answer: QuizQuestionOption, index: number) => {
-  const parts = (answer.answer_title || '').split('{dash}');
-
-  return parts.flatMap((part, partIndex) => {
-    const nodes = [];
-
-    if (part) {
-      nodes.push(<span key={`${answer.answer_id || index}-text-${partIndex}`}>{part}</span>);
-    }
-
-    if (partIndex < parts.length - 1) {
-      nodes.push(
-        <input
-          key={`${answer.answer_id || index}-input-${partIndex}`}
-          type="text"
-          className="tutor-quiz-question-input"
-          placeholder={__('Type your answer here', 'tutor')}
-          disabled
-        />,
-      );
-    }
-
-    return nodes;
-  });
 };
 
 const getPreviewFrameStyles = () => `
