+++ assets/src/js/v3/entries/course-builder/components/modals/question-previews/ScalePreview.tsx
@@ -1,0 +1,121 @@
+import { useEffect, useRef } from 'react';
+import { __ } from '@wordpress/i18n';
+import SVGIcon from '@TutorShared/atoms/SVGIcon';
+import { tutorConfig } from '@TutorShared/config/config';
+import { type QuizQuestionOption } from '@TutorShared/utils/types';
+
+const SCALE_SCRIPT_ATTR = 'data-tutor-scale-preview-script';
+
+const ScalePreview = ({ answers }: { answers: QuizQuestionOption[] }) => {
+  const wrapperRef = useRef<HTMLDivElement>(null);
+
+  const configRaw = answers[0]?.answer_two_gap_match ?? '';
+  let rawConfig: Record<string, unknown> = {};
+  try {
+    const parsed = JSON.parse(configRaw) as { config?: Record<string, unknown> };
+    rawConfig = parsed?.config ?? {};
+  } catch {
+    // use defaults
+  }
+
+  const min = (rawConfig.min as number) ?? 0;
+  const max = (rawConfig.max as number) ?? 100;
+  const step = (rawConfig.step as number) ?? 1;
+  const defaultValue = (rawConfig.defaultValue as number) ?? (min + max) / 2;
+  const pxPerUnit = (rawConfig.pxPerUnit as number) ?? 10;
+  const labelEvery = (rawConfig.labelEvery as number) ?? Math.max(1, (max - min) / 10);
+  const minorTickEvery = (rawConfig.minorTickEvery as number) ?? Math.max(1, (max - min) / 50);
+  const precision = (rawConfig.precision as number) ?? (step < 1 ? 2 : 0);
+
+  // Unique IDs matching scale-question.js selectors
+  const qId = answers[0]?.answer_id ?? 'preview';
+  const scaleConfig = JSON.stringify({
+    min,
+    max,
+    step,
+    defaultValue,
+    pxPerUnit,
+    labelEvery,
+    minorTickEvery,
+    precision,
+  });
+
+  useEffect(() => {
+    const wrapper = wrapperRef.current;
+    if (!wrapper) return;
+
+    const doc = wrapper.ownerDocument;
+    if (!doc || doc === document) return;
+
+    if (doc.head.querySelector(`[${SCALE_SCRIPT_ATTR}]`)) {
+      // Script already present — reset the wrapper init flag so it re-initializes
+      wrapper.removeAttribute('data-tutor-scale-init');
+      // Re-trigger initAll via a tiny inline script on the iframe window
+      const reinit = doc.createElement('script');
+      reinit.textContent = 'if(window._tutorScaleInitAll){window._tutorScaleInitAll();}';
+      doc.head.appendChild(reinit);
+      return;
+    }
+
+    const siteUrl = tutorConfig.site_url.replace(/\/$/, '');
+    const script = doc.createElement('script');
+    script.setAttribute(SCALE_SCRIPT_ATTR, '1');
+    // Patch initAll onto window so we can call it on re-renders
+    script.textContent = `
+      (function(){
+        var scriptEl = document.createElement('script');
+        scriptEl.src = '${siteUrl}/wp-content/plugins/tutor-pro/assets/js/scale-question.js';
+        scriptEl.onload = function(){
+          // Expose initAll for re-use on subsequent renders
+          if(typeof initAll === 'function'){ window._tutorScaleInitAll = function(){ initAll(false); }; }
+        };
+        document.head.appendChild(scriptEl);
+      })();
+    `;
+    doc.head.appendChild(script);
+  }, []);
+
+  return (
+    <div
+      ref={wrapperRef}
+      id={`tutor-scale-question-${qId}`}
+      className="tutor-quiz-question-options tutor-scale-question"
+      data-question-type="scale"
+      data-question-id={String(qId)}
+      data-scale-config={scaleConfig}
+    >
+      <div className="tutor-scale-slider-wrapper">
+        <div className="tutor-scale-bubble" id={`tutor-scale-bubble-${qId}`}>
+          <div className="tutor-scale-bubble-value">{defaultValue}</div>
+          <div className="tutor-scale-bubble-pointer" />
+        </div>
+        <div
+          className="tutor-scale-container"
+          id={`tutor-scale-container-${qId}`}
+          aria-label={__('Interactive scale: drag to select your answer value.', 'tutor')}
+        >
+          <div className="tutor-scale" id={`tutor-scale-${qId}`}>
+            {/* Ticks are generated by scale-question.js */}
+          </div>
+        </div>
+        <div className="tutor-scale-instructions">
+          <div className="tutor-scale-hand-icon" aria-hidden="true">
+            <SVGIcon name="handSwipeRight" width={20} height={20} />
+          </div>
+          <span className="tutor-scale-instruction-text">
+            {__('Move the scale left or right to set value', 'tutor')}
+          </span>
+        </div>
+      </div>
+      <input
+        type="hidden"
+        id={`tutor-scale-value-${qId}`}
+        name={`preview[answers][scale][value]`}
+        defaultValue=""
+        readOnly
+      />
+    </div>
+  );
+};
+
+export default ScalePreview;
