+++ assets/src/js/v3/entries/course-builder/components/modals/question-previews/UnsupportedPreview.tsx
@@ -1,0 +1,11 @@
+import { __ } from '@wordpress/i18n';
+
+const UnsupportedPreview = () => (
+  <div className="tutor-quiz-question-options">
+    <div className="tutor-quiz-question-option">
+      {__('Preview is not available for this question type yet.', 'tutor')}
+    </div>
+  </div>
+);
+
+export default UnsupportedPreview;
